requirements.txt
flask
flask_cors
flask-restx
flask-httpauth
waitress
psutil
requests_toolbelt
openai-whisper==20230314
wave
vosk
